cloud_function/requirements.txt
functions-framework
requests
pandas
google-cloud-bigquery
google-cloud-secret-manager
pyarrow